Fix fatal config parse panic on ability durations; bump SDK to 2.4.1
game-constants.json has had doubleDamageDurationSecs: 7.5 since April
(PR #75) while the Rust struct declared u64, so serde failed and every
first touch of shared_config::config() in a wasm instance panicked.
This broke client_connected on maincloud after today''s publishes and
made init crash on any fresh database (surfaced while standing up the
self-hosted dev SpacetimeDB on Coolify). Ability duration fields are
now f32 with Duration::from_secs_f32, honoring the JSON as the source
of truth. Also updates the SpacetimeDB SDK 2.0.3 -> 2.4.1 to match the
self-hosted server version.

+++ server/spacetimedb/src/constants.rs
@@ -578,7 +578,7 @@ pub fn ability_type_speed_boost() -> u8 {
 pub fn ability_type_shield() -> u8 {
     shared_config::config().abilities.types.shield
 }
-pub fn double_damage_duration_secs() -> u64 {
+pub fn double_damage_duration_secs() -> f32 {
     shared_config::config()
         .abilities
         .double_damage_duration_secs
@@ -586,13 +586,13 @@ pub fn double_damage_duration_secs() -> u64 {
 pub fn double_damage_multiplier() -> f32 {
     shared_config::config().abilities.double_damage_multiplier
 }
-pub fn speed_boost_duration_secs() -> u64 {
+pub fn speed_boost_duration_secs() -> f32 {
     shared_config::config().abilities.speed_boost_duration_secs
 }
 pub fn speed_boost_multiplier() -> f32 {
     shared_config::config().abilities.speed_boost_multiplier
 }
-pub fn shield_duration_secs() -> u64 {
+pub fn shield_duration_secs() -> f32 {
     shared_config::config().abilities.shield_duration_secs
 }
 pub fn shield_damage_reduction() -> f32 {
